# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rewrd API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Rewrd Engineering",
      "url": "https://getrewrd.com"
    }
  },
  "servers": [
    {
      "url": "https://rewrd-api-staging.fly.dev/v1",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:3004/v1",
      "description": "Local Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your API Key (e.g. sk_test_...)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all endpoints on failure.",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false,
            "description": "Always `false` for error responses."
          },
          "message": {
            "type": "string",
            "example": "Customer not found",
            "description": "Human-readable error message."
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "not_found",
                "description": "Machine-readable error code for programmatic handling."
              },
              "request_id": {
                "type": "string",
                "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                "description": "Unique identifier for this request, useful for support inquiries."
              },
              "description": {
                "type": "string",
                "example": "Customer not found",
                "description": "Detailed description of the error."
              }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "description": "Pagination metadata included in list responses.",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1,
            "description": "Current page number."
          },
          "limit": {
            "type": "integer",
            "example": 50,
            "description": "Number of items per page."
          },
          "total": {
            "type": "integer",
            "example": 150,
            "description": "Total number of items across all pages."
          },
          "total_pages": {
            "type": "integer",
            "example": 3,
            "description": "Total number of pages."
          }
        }
      },
...
      "PointsTransaction": {
        "type": "object",
        "description": "A ledger entry representing a points credit or redemption. Every point operation creates a transaction record.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for this transaction.",
            "example": 1284
          },
          "customer_uid": {
            "type": "string",
            "description": "The customer who earned or redeemed points.",
            "example": "26833813"
          },
          "points": {
            "type": "number",
            "description": "Number of points in this transaction.",
            "example": 50
          },
          "title": {
            "type": "string",
            "description": "Human-readable title for this transaction.",
            "example": "Points Credit"
          },
          "narration": {
            "type": "string",
            "description": "Optional description or note for this transaction.",
            "example": "Credited 50 points for purchase #12345"
          },
          "transaction_type": {
            "type": "string",
            "description": "Type of transaction:\n- `member_points_adjustment_credit` — direct point credit\n- `member_purchase_order_earned_fixed` — points earned via a fixed earning rule\n- `member_purchase_order_earned_percentage` — points earned via a percentage earning rule\n- `member_purchase_order_redeemed` — points redeemed\n",
            "example": "member_points_adjustment_credit"
          },
          "ledger_type": {
            "type": "string",
            "description": "Generic type of ledger entry (credit or debit).",
            "example": "credit"
          },
          "status": {
            "type": "string",
            "description": "Status of the transaction.",
            "example": "successful"
          },
          "reference_id": {
            "type": "string",
            "description": "Unique reference ID for this transaction, used for deduplication.",
            "example": "credit_a1b2c3d4-e5f6-7890"
          },
          "balance_before": {
            "type": "number",
            "description": "Customer's points balance before this transaction was applied.",
            "example": 1500
          },
          "balance_after": {
            "type": "number",
            "description": "Customer's points balance after this transaction was applied.",
            "example": 1550
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when this transaction was created.",
            "example": "2026-02-20T14:30:00.000Z"
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/points/transaction": {
      "post": {
        "summary": "Process a Unified Points Transaction",
        "operationId": "processPointsTransaction",
        "description": "Process a points transaction for a customer. This single endpoint handles three scenarios:\n\n**How it works:**\nUse the `reward` and `redeem` boolean flags to tell Rewrd what type of transaction to perform. You can set one or both to `true`.\n\n| `reward` | `redeem` | What happens |\n|---|---|---|\n| `true` | `false` | **Reward only** — Customer earns points based on an earning rule. |\n| `false` | `true` | **Redeem only** — Customer spends points toward a purchase. |\n| `true` | `true` | **Combined** — Customer earns AND redeems points in a single checkout. |\n\n---\n\n**When `reward` is `true`:**\nYou must provide `way_to_earn_id` (the ID of the earning rule from your dashboard). Rewrd will calculate the points to credit based on the rule's type:\n- **Fixed rule** — Awards a fixed number of points regardless of `order_value` (e.g. always award 50 points).\n- **Percentage rule** — Awards points as a percentage of `order_value` (e.g. 10% of ₦5,000 = 500 points).\n\n**When `redeem` is `true`:**\nYou must provide `deduct_points` — the exact number of points to deduct from the customer's balance. The customer must have enough points, otherwise the request will fail with an `insufficient_points` error.\n\n---\n\n**Real-world example — Combined checkout:**\nA customer buys a coffee for ₦5,000. They want to use 200 of their points AND earn new points for the purchase:\n```json\n{\n  \"customer_uid\": \"cust_abc123\",\n  \"order_id\": \"ORD-003\",\n  \"order_value\": 5000,\n  \"redeem\": true,\n  \"deduct_points\": 200,\n  \"reward\": true,\n  \"way_to_earn_id\": 1\n}\n```\nThis will deduct 200 points AND credit new points based on earning rule #1, all in a single atomic operation.\n\n---\n\nThis endpoint requires an `Idempotency-Key` header to prevent duplicate transactions. Use a unique key per checkout (e.g. your order ID).\n",
        "tags": [
          "Points"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$"
            },
            "description": "Unique key for idempotent request handling. Must be alphanumeric with hyphens or underscores, 1–100 characters.",
            "example": "tx-cust123-20260225-001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customer_uid",
                  "order_id",
                  "order_value",
                  "redeem",
                  "reward"
                ],
                "properties": {
                  "customer_uid": {
                    "type": "string",
                    "description": "The UID of the customer making the checkout.",
                    "example": "cust_abc123"
                  },
                  "order_id": {
                    "type": "string",
                    "description": "External order ID tying this points usage to a checkout basket.",
                    "example": "ORD-003"
                  },
                  "order_value": {
                    "type": "number",
                    "description": "The fiat value of the items in the total basket.",
                    "example": 10000
                  },
                  "redeem": {
                    "type": "boolean",
                    "description": "True if the customer is choosing to deduct points towards the purchase.",
                    "example": true
                  },
                  "reward": {
                    "type": "boolean",
                    "description": "True if the order creates an earning opportunity.",
                    "example": true
                  },
                  "deduct_points": {
                    "type": "integer",
                    "description": "Quantity of points exactly to be removed (required if redeem is true).",
                    "example": 200
                  },
                  "way_to_earn_id": {
                    "type": "integer",
                    "description": "Pre-configured loyalty mechanism integer rule (required if reward is true).",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction executed properly.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction processed successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PointsTransaction"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error — missing required condition triggers."
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key."
          },
          "404": {
            "description": "Customer or earning rule not found."
          }
        }
      }
    },
    "/points/customers/{uid}/transactions": {
      "get": {
        "summary": "Get Customer Transaction History",
        "operationId": "getCustomerTransactions",
        "description": "Retrieves a paginated list of all point transactions (credits and redemptions) for a specific customer.\n\nTransactions are sorted by date with the most recent first. Use this endpoint to display a customer's points activity history.\n\n**Understanding the response fields:**\n\n**`ledger_type`** tells you the direction of the transaction:\n- `credit` — Points were **added** to the customer's balance\n- `debit` — Points were **deducted** from the customer's balance\n\n**`transaction_type`** tells you *how* the points were earned or spent:\n| Transaction Type | Meaning |\n|---|---|\n| `member_points_adjustment_credit` | Points were directly credited (manual adjustment) |\n| `member_purchase_order_earned_fixed` | Points earned via a **fixed** earning rule (e.g. \"earn 50 points per purchase\") |\n| `member_purchase_order_earned_percentage` | Points earned via a **percentage** earning rule (e.g. \"earn 10% of order value as points\") |\n| `member_purchase_order_redeemed` | Points were redeemed (spent) toward a purchase |\n\n**`balance_before`** and **`balance_after`** show the customer's points balance before and after the transaction, so you can verify the math.\n",
        "tags": [
          "Points"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the customer.",
            "example": "26833813"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination (1-indexed).",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Number of transactions to return per page.",
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transactions retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PointsTransaction"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}